# Backend/api/laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Proyecto Final Grado API",
        "description": "Documentacion de la API para el Proyecto Final de Grado",
        "contact": {
            "name": "Alejandro Solar, David Kovacs",
            "email": "alejandrosolarruiz@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/noticias": {
            "get": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Listar todas las noticias paginadas (solo listado)",
                "operationId": "914e63ade6e06a3fc689556f6c295245",
                "parameters": [
                    {
                        "name": "pagina",
                        "in": "query",
                        "description": "Número de página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "registrosPorPagina",
                        "in": "query",
                        "description": "Cantidad de registros por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de noticias (solo listado, sin metadatos de paginación)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Noticia"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Crear una nueva noticia",
                "operationId": "8f1e4f67dea603a29aa7c4817f7810ea",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NoticiaCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Noticia creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Noticia"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/noticias/{id}": {
            "get": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Obtener una noticia por ID",
                "operationId": "ce8296da4c297f04ae5dd0493b9c28b0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la noticia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Noticia encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Noticia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Noticia no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Actualizar una noticia existente",
                "operationId": "5b09bf8daeef7b4b31b386dd7d814079",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la noticia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NoticiaUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Noticia actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Noticia"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Noticia no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Noticia"
                ],
                "summary": "Eliminar una noticia",
                "operationId": "fdd8118e7eb14c73fa335c678da6b85a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la noticia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Noticia eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Noticia eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Noticia no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Noticia": {
                "properties": {
                    "id_noticia": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Título de la noticia"
                    },
                    "imagen": {
                        "description": "Imagen en formato base64",
                        "type": "string",
                        "format": "byte",
                        "nullable": true
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción corta"
                    },
                    "cuerpo": {
                        "type": "string",
                        "example": "Texto completo de la noticia"
                    }
                },
                "type": "object"
            },
            "NoticiaCreate": {
                "required": [
                    "titulo",
                    "descripcion",
                    "cuerpo"
                ],
                "properties": {
                    "titulo": {
                        "type": "string",
                        "example": "Título de la noticia"
                    },
                    "imagen": {
                        "description": "Imagen en formato base64",
                        "type": "string",
                        "format": "byte",
                        "nullable": true
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción corta"
                    },
                    "cuerpo": {
                        "type": "string",
                        "example": "Texto completo de la noticia"
                    }
                },
                "type": "object"
            },
            "NoticiaUpdate": {
                "properties": {
                    "titulo": {
                        "type": "string",
                        "example": "Título de la noticia"
                    },
                    "imagen": {
                        "description": "Imagen en formato base64",
                        "type": "string",
                        "format": "byte",
                        "nullable": true
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción corta"
                    },
                    "cuerpo": {
                        "type": "string",
                        "example": "Texto completo de la noticia"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}